Migrate scheduleControlService to TypeScript
